pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "icarecommunication"
version = "0.1.0"
dependencies = [
    "paho-mqtt>=2.0.0"
]
requires-python = ">=3.10"